+++ src/extension/Entities/accessModifier.ts
@@ -1,0 +1,6 @@
+export enum AccessModifier {
+    local = 'local ',
+    protected = 'protected ',
+    internal = 'internal ',
+    public = ''
+}