+++ src/test/suite/ALExtractProcedureCA.test.ts
@@ -3,14 +3,15 @@ import * as path from 'path';
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
 import * as vscode from 'vscode';
-import { CodeActionProviderExtractProcedure } from '../../extension/Services/CodeActionProviderExtractProcedure';
-import { ALProcedure } from '../../extension/Entities/alProcedure';
-import { ReturnTypeAnalyzer } from '../../extension/Extract Procedure/returnTypeAnalyzer';
-import { ALLanguageExtension } from '../alExtension';
-import { ALTestProject } from './ALTestProject';
+import { FullSyntaxTreeNodeKind } from '../../extension/AL Code Outline Ext/fullSyntaxTreeNodeKind';
 import { ALFullSyntaxTreeNode } from '../../extension/AL Code Outline/alFullSyntaxTreeNode';
 import { SyntaxTree } from '../../extension/AL Code Outline/syntaxTree';
-import { FullSyntaxTreeNodeKind } from '../../extension/AL Code Outline Ext/fullSyntaxTreeNodeKind';
+import { AccessModifier } from '../../extension/Entities/accessModifier';
+import { ALProcedure } from '../../extension/Entities/alProcedure';
+import { ReturnTypeAnalyzer } from '../../extension/Extract Procedure/returnTypeAnalyzer';
+import { CodeActionProviderExtractProcedure } from '../../extension/Services/CodeActionProviderExtractProcedure';
+import { ALLanguageExtension } from '../alExtension';
+import { ALTestProject } from './ALTestProject';
 
 
 suite('ALExtractProcedureCA Test Suite', function () {
@@ -50,7 +51,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'integer');
@@ -74,7 +75,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         //current solution because everything is handed over as var
         assert.strictEqual(alProcedure.parameters.length, 1);
@@ -100,7 +101,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         //current solution because everything is handed over as var
         assert.strictEqual(alProcedure.parameters.length, 1);
@@ -126,7 +127,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         //current solution because everything is handed over as var
         assert.strictEqual(alProcedure.parameters.length, 1);
@@ -152,7 +153,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer');
@@ -172,7 +173,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer');
@@ -192,7 +193,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer');
@@ -213,7 +214,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'codeunit codeunittoextract');
@@ -234,7 +235,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'page mypage');
@@ -255,7 +256,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer');
@@ -278,7 +279,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer');
@@ -298,7 +299,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         //current solution because everything is handed over as var
         assert.strictEqual(alProcedure.parameters.length, 1);
@@ -324,7 +325,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer');
@@ -346,7 +347,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 1);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer');
@@ -366,7 +367,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.parameters.length, 2);
         assert.strictEqual(alProcedure.parameters[0].type.toLowerCase(), 'record customer temporary');
@@ -389,7 +390,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, 'Boolean');
         assert.strictEqual(alProcedure.returnVariableName, 'returnValue');
         assert.strictEqual(alProcedure.getBody(), 'returnValue := 1 in [2, 3];')
@@ -408,7 +409,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunitToExtractDocument, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, 'Integer');
         assert.strictEqual(alProcedure.returnVariableName, 'returnValue');
         assert.strictEqual(alProcedure.getBody(), 'returnValue := (addend - 1);')
@@ -431,7 +432,7 @@ suite('ALExtractProcedureCA Test Suite', function () {
         let alProcedure: ALProcedure | undefined = await new CodeActionProviderExtractProcedure(codeunit1Document, rangeToExtract).provideProcedureObjectForCodeAction(rangeToExtract, returnTypeAnalyzer);
         assert.notStrictEqual(alProcedure, undefined, 'Procedure should be extracted');
         alProcedure = alProcedure as ALProcedure;
-        assert.strictEqual(alProcedure.isLocal, true);
+        assert.strictEqual(alProcedure.accessModifier, AccessModifier.local);
         assert.strictEqual(alProcedure.returnType, undefined);
         assert.strictEqual(alProcedure.getBody(), 'intParam := 6;')
         assert.strictEqual(alProcedure.parameters.length, 1);