+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to the "al-codeactions" extension will be documented in this
 
 ## 0.2.22
 
-- Extract procedure: fixed if one-liner[#97](https://github.com/DavidFeldhoff/al-codeactions/issues/97)
+- Extract procedure:
+  - fixed if one-liner[#97](https://github.com/DavidFeldhoff/al-codeactions/issues/97)
+  - Create Event Publishers in other documents than the current one. [#99](https://github.com/DavidFeldhoff/al-codeactions/issues/99)
 - New functionality: Fix Cop Warnings
   - Remove assigned, but unused variables (safe and unsafe mode)
   - Remove unused variables